refactor: Improve payment amount handling and add payment validation logic
- Update amount conversion to use direct casting instead of rounding
- Modify payment creation services to validate existing payments
- Add checks for pending status and matching amount in payment reuse
- Enhance logging for payment creation and validation scenarios
- Update constructor signatures to include GetPaymentInterface service

+++ Controller/Payment/Redirect.php
@@ -75,7 +75,7 @@ class Redirect implements HttpGetActionInterface
         $order = $this->checkoutSession->getLastRealOrder();
 
         $data = [
-            'amount' => $order->getBaseGrandTotal() * 100,
+            'amount' => (int) ($order->getBaseGrandTotal() * 100),
             'currency' => $order->getBaseCurrencyCode(),
             'orderId' => $order->getIncrementId(),
             'customer' => $this->getCustomerDetails($order),
